General: Move geocoding tests after the actual classes
